requirements.txt
PyMuPDF>=1.23
pypdf>=4.0
pdfplumber>=0.11
PyYAML>=6.0
numpy>=1.24
scikit-learn>=1.3
